refactor: Update examples directory and readme
