+++ examples/rgbpp/spore/6-unlock-btc-time-cell.ts
@@ -1,5 +1,5 @@
 import { sendCkbTx, getBtcTimeLockScript, buildSporeBtcTimeCellsSpentTx, signBtcTimeCellSpentTx } from '@rgbpp-sdk/ckb';
-import { CKB_PRIVATE_KEY, btcService, ckbAddress, collector, isMainnet } from '../core';
+import { CKB_PRIVATE_KEY, btcService, ckbAddress, collector, isMainnet } from '../utils';
 
 // Warning: Wait at least 6 BTC confirmation blocks to spend the BTC time cells after 4-btc-jump-ckb.ts
 const unlockSporeBtcTimeCell = async ({ btcTimeCellArgs }: { btcTimeCellArgs: string }) => {
