+++ examples/rgbpp/spore/launch/1-prepare-cluster.ts
@@ -10,7 +10,7 @@ import {
   genRgbppLockScript,
   getSecp256k1CellDep,
 } from '@rgbpp-sdk/ckb';
-import { ckbAddress, isMainnet, collector, CKB_PRIVATE_KEY } from '../../core';
+import { ckbAddress, isMainnet, collector, CKB_PRIVATE_KEY } from '../../utils';
 import { CLUSTER_DATA } from './0-cluster-info';
 
 const prepareClusterCell = async ({ outIndex, btcTxId }: { outIndex: number; btcTxId: string }) => {
