--- examples/rgbpp/tsconfig.build.json
@@ -1,10 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "module": "CommonJS",
-    "rootDir": "src",
-    "outDir": "lib",
-    "noEmit": false
-  },
-  "include": ["core/*.ts"],
-}
