+++ examples/rgbpp/utils/index.ts
