+++ examples/rgbpp/utils/xudt.ts
@@ -1,67 +1,7 @@
 import { serializeScript } from '@nervosnetwork/ckb-sdk-utils';
-import {
-  genCkbJumpBtcVirtualTx,
-  getSecp256k1CellDep,
-  buildRgbppLockArgs,
-  getXudtTypeScript,
-  genBtcTransferCkbVirtualTx,
-  genBtcJumpCkbVirtualTx,
-} from '@rgbpp-sdk/ckb';
-import {
-  CKB_PRIVATE_KEY,
-  isMainnet,
-  collector,
-  ckbAddress,
-  btcAddress,
-  btcKeyPair,
-  btcService,
-  btcDataSource,
-} from './env';
+import { getXudtTypeScript, genBtcTransferCkbVirtualTx, genBtcJumpCkbVirtualTx } from '@rgbpp-sdk/ckb';
+import { isMainnet, collector, btcAddress, btcKeyPair, btcService, btcDataSource } from './env';
 import { sendRgbppUtxos } from '@rgbpp-sdk/btc';
-
-export interface LeapToBtcParams {
-  outIndex: number;
-  btcTxId: string;
-  xudtTypeArgs: string;
-  transferAmount: bigint;
-}
-
-export const leapFromCkbToBtc = async ({
-  outIndex,
-  btcTxId,
-  xudtTypeArgs,
-  transferAmount,
-}: LeapToBtcParams): Promise<string> => {
-  const toRgbppLockArgs = buildRgbppLockArgs(outIndex, btcTxId);
-
-  // Warning: Please replace with your real xUDT type script here
-  const xudtType: CKBComponents.Script = {
-    ...getXudtTypeScript(isMainnet),
-    args: xudtTypeArgs,
-  };
-
-  const ckbRawTx = await genCkbJumpBtcVirtualTx({
-    collector,
-    fromCkbAddress: ckbAddress,
-    toRgbppLockArgs,
-    xudtTypeBytes: serializeScript(xudtType),
-    transferAmount,
-  });
-
-  const emptyWitness = { lock: '', inputType: '', outputType: '' };
-  const unsignedTx: CKBComponents.RawTransactionToSign = {
-    ...ckbRawTx,
-    cellDeps: [...ckbRawTx.cellDeps, getSecp256k1CellDep(false)],
-    witnesses: [emptyWitness, ...ckbRawTx.witnesses.slice(1)],
-  };
-
-  const signedTx = collector.getCkb().signTransaction(CKB_PRIVATE_KEY)(unsignedTx);
-
-  const txHash = await collector.getCkb().rpc.sendTransaction(signedTx, 'passthrough');
-  console.info(`Rgbpp asset has been jumped from CKB to BTC and tx hash is ${txHash}`);
-
-  return txHash;
-};
 
 export interface RgbppTransferParams {
   rgbppLockArgsList: string[];
