+++ examples/xudt/env.ts
@@ -2,11 +2,9 @@ import { AddressPrefix, privateKeyToAddress } from '@nervosnetwork/ckb-sdk-utils
 import { Collector } from '@rgbpp-sdk/ckb';
 import dotenv from 'dotenv';
 
-dotenv.config({ path: __dirname + '/../.env' });
+dotenv.config({ path: __dirname + '/.env' });
 
 export const isMainnet = process.env.IS_MAINNET === 'true' ? true : false;
-
-console.log('IS_MAINNET', process.env.IS_MAINNET);
 
 export const collector = new Collector({
   ckbNodeUrl: process.env.CKB_NODE_URL!,
