+++ examples/xudt/package.json
@@ -5,7 +5,6 @@
   "private": true,
   "type": "commonjs",
   "scripts": {
-    "build": "tsc -p tsconfig.build.json",
     "format": "prettier --write '**/*.ts'",
     "lint": "tsc && eslint . && prettier --check '**/*.ts'",
     "lint:fix": "tsc && eslint --fix --ext .ts . && prettier --write '**/*.ts'"
